requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
jinja2
aiofiles
python-dotenv
psutil
pydantic-settings
requests
